feat: rescue LoginShellProbe and PathShimInstaller into Capacitor.Cli.Core
Move the process seam (IProcessRunner + records + the production ProcessRunner
implementation) and the two Avalonia-only setup classes into Core so the flow
can drive them after AI-2053 deletes the app. App consumers pick them up via
using; the installer's destination-override seam becomes public (Core has no
InternalsVisibleTo for the app). Tests move with the classes.
