+++ src/Capacitor.App/Services/ShimOfferCoordinator.cs
@@ -1,4 +1,5 @@
 using System.Reactive.Subjects;
+using Capacitor.Cli.Core.Setup;
 
 namespace Capacitor.App.Services;
