+++ src/Capacitor.App/ViewModels/Onboarding/ImportStepViewModel.cs
@@ -2,6 +2,7 @@ using System.Collections.ObjectModel;
 using System.Reactive;
 using System.Reactive.Linq;
 using Capacitor.App.Services;
+using Capacitor.Cli.Core;
 using Capacitor.Cli.Core.Setup;
 using ReactiveUI;
