+++ test/Capacitor.App.Tests.Unit/AppMutationLaneWiringTests.cs
@@ -1,5 +1,6 @@
 using Capacitor.App.Services;
 using Capacitor.App.Services.Mutation;
+using Capacitor.Cli.Core;
 using AppUnderTest = Capacitor.App.App;
 
 namespace Capacitor.App.Tests.Unit;
