+++ test/Capacitor.App.Tests.Unit/DaemonMutationLaneTests.cs
@@ -1,5 +1,7 @@
 using Capacitor.App.Services;
 using Capacitor.App.Services.Mutation;
+using Capacitor.Cli.Core;
+using Capacitor.Cli.Core.Setup;
 using Microsoft.Extensions.Time.Testing;
 
 namespace Capacitor.App.Tests.Unit;
