+++ test/Capacitor.App.Tests.Unit/WizardSimpleStepsTests.cs
@@ -7,7 +7,9 @@ using Avalonia.VisualTree;
 using Capacitor.App.Services;
 using Capacitor.App.ViewModels.Onboarding;
 using Capacitor.App.Views.Onboarding;
+using Capacitor.Cli.Core;
 using Capacitor.Cli.Core.Config;
+using Capacitor.Cli.Core.Setup;
 using TUnit.Assertions.Enums;
 
 namespace Capacitor.App.Tests.Unit;
