+++ test/Capacitor.Cli.Core.Tests.Unit/PathShimInstallerTests.cs
@@ -1,7 +1,7 @@
 using System.Diagnostics;
-using Capacitor.App.Services;
+using Capacitor.Cli.Core.Setup;
 
-namespace Capacitor.App.Tests.Unit;
+namespace Capacitor.Cli.Core.Tests.Unit;
 
 public class PathShimInstallerTests {
     sealed class FakeProcessRunner : IProcessRunner {
